+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/CheckEntityAction.java
@@ -1,5 +1,6 @@
 package com.elmakers.mine.bukkit.action.builtin;
 
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -18,6 +19,8 @@ public class CheckEntityAction extends CheckAction {
     private Boolean onFire;
     private Set<EntityType> allowedTypes;
     private Set<EntityType> deniedTypes;
+    private List<Class<? extends Entity>> allowedClasses;
+    private List<Class<? extends Entity>> deniedClasses;
 
     @Override
     public void prepare(CastContext context, ConfigurationSection parameters)
@@ -53,6 +56,43 @@ public class CheckEntityAction extends CheckAction {
                 }
             }
         }
+
+        if (parameters.contains("allowed_entity_classes")) {
+            List<String> keys = ConfigurationUtils.getStringList(parameters, "allowed_entity_classes");
+            allowedClasses = new ArrayList<>();
+            for (String key : keys) {
+                try {
+                    Class<?> rawClass = Class.forName("org.bukkit.entity." + key);
+                    if (!Entity.class.isAssignableFrom(rawClass)) {
+                        context.getLogger().warning("Invalid entity class in CheckEntity configuration, does not extend from Entity: " + rawClass.getName());
+                        continue;
+                    }
+                    @SuppressWarnings("unchecked")
+                    Class<? extends Entity> targetEntityType = (Class<? extends Entity>)rawClass;
+                    allowedClasses.add(targetEntityType);
+                } catch (Exception ex) {
+                    context.getLogger().warning("Invalid entity class in CheckEntity configuration: " + key);
+                }
+            }
+        }
+        if (parameters.contains("denied_entity_classes")) {
+            List<String> keys = ConfigurationUtils.getStringList(parameters, "denied_entity_classes");
+            deniedClasses = new ArrayList<>();
+            for (String key : keys) {
+                try {
+                    Class<?> rawClass = Class.forName("org.bukkit.entity." + key);
+                    if (!Entity.class.isAssignableFrom(rawClass)) {
+                        context.getLogger().warning("Invalid entity class in CheckEntity configuration, does not extend from Entity: " + rawClass.getName());
+                        continue;
+                    }
+                    @SuppressWarnings("unchecked")
+                    Class<? extends Entity> targetEntityType = (Class<? extends Entity>)rawClass;
+                    deniedClasses.add(targetEntityType);
+                } catch (Exception ex) {
+                    context.getLogger().warning("Invalid entity class in CheckEntity configuration: " + key);
+                }
+            }
+        }
     }
 
     @Override
@@ -66,16 +106,33 @@ public class CheckEntityAction extends CheckAction {
         if (onlyCaster && !isCaster) {
             return false;
         }
-        if (allowedTypes != null && !allowedTypes.contains(targetEntity.getType())) {
+        if (onFire != null && onFire != (targetEntity.getFireTicks() > 0)) {
             return false;
         }
         if (deniedTypes != null && deniedTypes.contains(targetEntity.getType())) {
             return false;
         }
-        if (onFire != null && onFire != (targetEntity.getFireTicks() > 0)) {
-            return false;
+        if (deniedClasses != null) {
+            for (Class<? extends Entity> entityClass : deniedClasses) {
+                if (entityClass.isAssignableFrom(targetEntity.getClass())) {
+                    return false;
+                }
+            }
         }
-        return true;
+
+        boolean anyAllowed = allowedTypes == null && allowedClasses == null;
+        if (!anyAllowed && allowedTypes != null) {
+            anyAllowed = allowedTypes.contains(targetEntity.getType());
+        }
+        if (!anyAllowed && allowedClasses != null) {
+            for (Class<? extends Entity> entityClass : allowedClasses) {
+                if (entityClass.isAssignableFrom(targetEntity.getClass())) {
+                    anyAllowed = true;
+                    break;
+                }
+            }
+        }
+        return anyAllowed;
     }
 
     @Override